tests: add setup/teardown for asok dir

+++ src/vstart.sh
@@ -343,7 +343,9 @@ else
     if [ "$new" -ne 0 ]; then
         # only delete if -n
         asok_dir=`dirname $($CEPH_BIN/ceph-conf --show-config-value admin_socket)`
-        [ -d $asok_dir ] && rm -f $asok_dir/* && rmdir $asok_dir
+        if [ $asok_dir != /var/run/ceph ]; then
+            [ -d $asok_dir ] && rm -f $asok_dir/* && rmdir $asok_dir
+        fi
         if [ -z "$CEPH_ASOK_DIR" ]; then
             CEPH_ASOK_DIR=`mktemp -u -d "${TMPDIR:-/tmp}/ceph-asok.XXXXXX"`
         fi
